created the middleware for authorization

+++ src/app/api/admin/delete-product/route.js
@@ -1,4 +1,5 @@
 import connectToDB from "@/database";
+import AuthUser from "@/middleware/AuthUser";
 import Product from "@/models/product";
 import { NextResponse } from "next/server";
 
@@ -7,28 +8,36 @@ export const dynamic = "force-dynamic";
 export async function DELETE(req) {
   try {
     await connectToDB();
-    //fetch the url and get the id of the post we want to delete
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
+    const isAuthUser = await AuthUser(req);
+    if (isAuthUser?.role === "admin") {
+      //fetch the url and get the id of the post we want to delete
+      const { searchParams } = new URL(req.url);
+      const id = searchParams.get("id");
 
-    if (!id) {
-      return NextResponse.json({
-        success: false,
-        message: "No such product exist!",
-      });
-    }
+      if (!id) {
+        return NextResponse.json({
+          success: false,
+          message: "No such product exist!",
+        });
+      }
 
-    const deletedProduct = await Product.findByIdAndDelete(id);
+      const deletedProduct = await Product.findByIdAndDelete(id);
 
-    if (deletedProduct) {
-      return NextResponse.json({
-        success: true,
-        message: "Product deleted successfully!",
-      });
+      if (deletedProduct) {
+        return NextResponse.json({
+          success: true,
+          message: "Product deleted successfully!",
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: "Couldn't delete the product, please try again later!",
+        });
+      }
     } else {
       return NextResponse.json({
         success: false,
-        message: "Couldn't delete the product, please try again later!",
+        message: "You are not authorized!",
       });
     }
   } catch (e) {
